requirements.txt
ffmpeg-python
moviepy 
pydub
